refactoring and bug fixes

+++ src/user-definitions/user-definitions-service.js
@@ -1,13 +1,28 @@
 const UserDefinitionsService = {
-  getAllDefinitions(db) {
-    return db.select('*').from('saved_definition')
+  getAllDefinitions(db, user_id) {
+    return db
+      .select('*')
+      .from('saved_definition')
+      .where({ user_id })
+      .join('definition', { 'definition.id': 'saved_definition.definition_id' })
   },
-  getByDefinitionId(db, definition_id) {
+  getByDefinitionId(db, user_id, definition_id) {
     return db
       .from('saved_definition')
       .select('*')
-      .where({ definition_id })
+      .where({ definition_id, user_id })
       .first()
+      .join('definition', { 'definition.id': 'saved_definition.definition_id' })
+  },
+  getByWordId(db, user_id, word_id) {
+    return db
+      .from('saved_definition')
+      .select('*')
+      .where({ user_id })
+      .join('definition', {
+        'definition.id': 'saved_definition.definition_id',
+        word_id: word_id,
+      })
   },
   insertDefinition(db, newDefinition) {
     return db
@@ -18,8 +33,8 @@ const UserDefinitionsService = {
         return rows[0]
       })
   },
-  deleteDefinition(db, definition_id) {
-    return db('saved_definition').where({ definition_id }).delete()
+  deleteDefinition(db, user_id, definition_id) {
+    return db('saved_definition').where({ definition_id, user_id }).delete()
   },
 }
